@@ -43,7 +43,7 @@ class _TemplateResponse(HTMLResponse):
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         request = self.context.get("request", {})
         extensions = request.get("extensions", {})
-        if "http.response.debug" in extensions:
+        if "http.response.debug" in extensions:  # pragma: no branch
             await send(
                 {
                     "type": "http.response.debug",
@@ -98,7 +98,7 @@ class Jinja2Templates:
         self.context_processors = context_processors or []
         if directory is not None:
             self.env = self._create_env(directory, **env_options)
-        elif env is not None:
+        elif env is not None:  # pragma: no branch
             self.env = env
 
         self._setup_env_defaults(self.env)
@@ -168,9 +168,9 @@ class Jinja2Templates:
                 name = args[0]
                 context = args[1] if len(args) > 1 else kwargs.get("context", {})
                 status_code = args[2] if len(args) > 2 else kwargs.get("status_code", 200)
-                headers = args[2] if len(args) > 2 else kwargs.get("headers")
-                media_type = args[3] if len(args) > 3 else kwargs.get("media_type")
-                background = args[4] if len(args) > 4 else kwargs.get("background")
+                headers = args[3] if len(args) > 3 else kwargs.get("headers")
+                media_type = args[4] if len(args) > 4 else kwargs.get("media_type")
+                background = args[5] if len(args) > 5 else kwargs.get("background")
 
                 if "request" not in context:
                     raise ValueError('context must include a "request" key')
