@@ -111,7 +111,7 @@ class WebSocket(HTTPConnection):
     ) -> None:
         headers = headers or []
 
-        if self.client_state == WebSocketState.CONNECTING:
+        if self.client_state == WebSocketState.CONNECTING:  # pragma: no branch
             # If we haven't yet seen the 'connect' message, then wait for it first.
             await self.receive()
         await self.send({"type": "websocket.accept", "subprotocol": subprotocol, "headers": headers})
